requirements.txt
# 核心依赖
biopython>=1.83
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# AI/深度学习
torch>=2.0.0
transformers>=4.35.0
fair-esm>=2.0.0
alphabetify>=0.1.0

# 结构预测
alphafold-colabfold>=1.5.0  # 轻量级版本
py3Dmol>=2.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0

# 序列分析
cutadapt>=4.6.0
seqtk>=1.3.0  # 需要单独安装

# 遗传密码表
genetic>=1.2.0

# 工具
tqdm>=4.66.0
pyyaml>=6.0.0
click>=8.1.0

# 数据库API
bioservices>=1.11.0
mygene>=3.2.2
mychem>=1.0.2
